ml/requirements.txt
concrete-ml
numpy
pandas
scikit-learn
fastapi
uvicorn
pydantic
python-multipart
hashlib
